requirements.txt
# Production Requirements - Complete
# All dependencies for advanced production deployment

# Core Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# WebSocket Support
flask-socketio==5.3.6
python-socketio==5.11.0

# Security & Authentication
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
pyjwt==2.8.0
cryptography==41.0.7

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Caching & Async
redis==5.0.1

# Validation
pydantic==2.5.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.3.0

# HTTP & Networking
requests==2.31.0
urllib3==2.1.0

# Scientific Computing (for FHE)
numpy==1.26.2

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0
